requirements.txt
tensorflow_addons==0.11.2
sklearn
matplotlib
pandas



